Apply rule RUF046

+++ tests/tensor/test_elemwise.py
@@ -255,7 +255,7 @@ class TestBroadcast:
     cop = Elemwise
 
     openmp_minsize = 2 * config.openmp_elemwise_minsize
-    openmp_minsize_sqrt = int(math.ceil(math.sqrt(openmp_minsize)))
+    openmp_minsize_sqrt = math.ceil(math.sqrt(openmp_minsize))
 
     # The order is important if you change them.
     linkers = [PerformLinker, CLinker]
